refactor: move JSON parsers to network package
